Update engine and use local locking for terrain modifications

+++ ScapesCore/src/main/kotlin/org/tobi29/scapes/entity/client/MobLivingClient.kt
@@ -101,7 +101,7 @@ abstract class MobLivingClient(type: EntityType<*, *>,
                 if (footStepSound != null) {
                     val random = threadLocalRandom()
                     world.playSound(footStepSound, this,
-                            0.9f + random.nextFloat() * 0.2f, 1.0f)
+                            0.9 + random.nextDouble() * 0.2, 1.0)
                     footStep = 1.0 / clamp(speed.now().length(), 1.0, 4.0)
                 }
             }
